chore: unify API startup script
