+++ scripts/setup_env.py
@@ -53,7 +53,7 @@ def main() -> None:
     if validate_configuration():
         print("Next steps:")
         print(" 1. Review the generated .env file and fill in required values.")
-        print(" 2. Start the application with `python start_api.py`.")
+        print(" 2. Start the application with `python start_api.py` (unified startup).")
     else:
         print("Please resolve configuration errors before starting the application.")
