--- start_api.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-echo "🚀 Starting Yosai Intel Dashboard API..."
-python3 start_api.py
